+++ docs/DeveloperGuide.adoc
@@ -258,9 +258,32 @@ Classes used by multiple components are in the `planmysem.commons` package.
 
 This section describes some noteworthy details on how certain features are implemented.
 
-=== Initializing of the Planner and it's Semesters
+=== Initialization of the Planner and it's Semester
 
-==== Future Implementation
+The Planner and it's Semester has to be initialized for _PlanMySem_ to work as all other features of _PlanMySem_ would
+interact with this Semester object.
+
+==== Current Implementation
+
+Upon launching _PlanMySem_, the initialization of the Planner and it's Semester would be implemented via two steps:
+
+1. Automatically generate the academic calendar from the current date.
+2. Setup current Semester from the academic calendar.
+
+//The academic calendar is dynamically generated by invoking the function
+//based on the current date retrieved from the system clock.
+//TBC
+==== Design Considerations
+
+===== Aspect: Generation of academic calendar
+
+* *Alternative 1 (current choice):* Generate academic calendar by performing calculations from the current date.
+** Pros: Generation of academic calendar is dynamic and will work for future dates.
+** Cons: Computationally expensive as many operations have to be performed.
+
+* *Alternative 2:* Retrieve academic calendar from a pre-generated file.
+** Pros: Generation of academic calendar is efficient and not prone to calculation errors.
+** Cons: Requires the pre-generated file which may be accidentally edited or deleted by the user.
 
 === Reinventing the Parser / Command Format and Structure
